bug fixed

+++ app/src/main/java/com/example/fooddeliveryproject/Models/OrderedFood.kt
@@ -6,7 +6,7 @@ data class OrderedFood(
     var name: String="",
     var description: String="",
     var imageUrl: String="",
-    var price: Int=0,
+    var price: Double=0.0,
     var category:String="",
     var soldCount:Int?=null,
     var orderedDate:String?=null,
